+++ api/api/src/main/java/zelenaLipa/api/rowMappers/RoleRowMapper.java
@@ -1,7 +1,7 @@
 package zelenaLipa.api.rowMappers;
 
 import org.springframework.jdbc.core.RowMapper;
-import zelenaLipa.api.rows.Role;
+import zelenaLipa.api.domain.Role;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -13,7 +13,7 @@ public class RoleRowMapper implements RowMapper<Role> {
 
         Role role = new Role();
 
-        role.setRoleId(rs.getLong("roleid"));
+        role.setRoleId(rs.getInt("roleid"));
         role.setName(rs.getString("name"));
 
         return role;
